添加入住时间判断，取消订单时间判断，

+++ api_1_0/orderFormResource/orderFormOtherResource.py
@@ -2,16 +2,17 @@
 # -*- coding:utf-8 -*-
 import datetime
 
-from werkzeug.exceptions import BadRequest
-from flask_restful import Resource, reqparse
-import flask
 from flask import jsonify
+from flask_restful import Resource, reqparse
+from werkzeug.exceptions import BadRequest
 
 from middlewares.JwtMiddleware import TokenRequire
 from service.orderFormService import OrderFormService
+from service.roomService import RoomService
+from utils.calc_time import Calculate
 from utils.myLogging import logger
-from utils.response_code import RET
 from utils.responseParser import ResponseParser
+from utils.response_code import RET
 
 
 class GetUserOrderFormInfoResource(Resource):
@@ -43,6 +44,21 @@ class CancelOrderFormResource(Resource):
         parser.add_argument('OrderFormID', type=int, location='json', required=True)
         try:
             data = parser.parse_args()
+            res = OrderFormService.get(OrderFormID=data.get('OrderFormID'))
+            if res.get('code') != RET.OK:
+                logger.error(res.get('data').get('error'))
+                return jsonify(ResponseParser.parse_res(**res))
+            orders = res.get('data')
+            for order in orders:
+                if Calculate.calc_time_diff_days(datetime.datetime.now(), order.get('ArrivalTime')) < 0:
+                    return jsonify({
+                        "code": RET.INTERNALERR,
+                        "message": "订单以逾期",
+                    })
+                res = RoomService.update(RoomID=order.get('RoomID'), RoomStatus=0)
+                if res.get('code') != RET.OK:
+                    logger.error(res.get('data').get('error'))
+                    return jsonify(ResponseParser.parse_res(**res))
             res = OrderFormService.delete(OrderFormID=data.get('OrderFormID'))
             if res.get("code") != RET.OK:
                 logger.error(res.get("data").get("error"))
@@ -54,5 +70,3 @@ class CancelOrderFormResource(Resource):
             return jsonify(ResponseParser.parse_param_error(error=str(e)))
         except Exception as e:
             return jsonify(ResponseParser.parse_unknown_error(error=str(e)))
-
-
